Major refactor: Use channels.
This resolves most of the problems of the initial implementation.
Instead of exposing handler traits, the main protocol now has to be
looped by the implementor. Each channel opened is Send, so it can be
transfered into a task, where it can be looped.

It's easy to add back some helper traits.